Simplify handleLink() to compare $text to $target->getPrefixedTitle() and $target->getTitle() instead of attempting using Title::newFromText() to see if user supplied custom link text.
This simplification enables the use of DisplayTitle for Categories.

Added Categories to unit tests.

Categories have slightly different syntax then standard links and
category links are generated by OutputPage.php

As it is lower case first letter doesn't override DisplayTitle for
categories because they correct their case before requesting a
a link to be rendered. This may or may not be overcommable by
implementing onOutputPageMakeCategoryLinks() in
DisplayTitleHooks.php.

If anyone ever makes this change, they will have to change the
logic in the unit test that ignores lower case if category.

Change-Id: I160b8721b07e2be4d807413cd86b0e176014c6be