Add config option for changing redirect display
There's an outstanding issue with the current implementation where the
updated display title applies to EVERY link on the page, including the
text that says "redirected from" at the top of a page when you are
redirected from somewhere else. I'm not sure how to fix that, so I'm not
addressing that issue in this patch; however, I do want the option to
disable the redirect-following-title-display behavior regardless of that
issue.

Bug: T217182
Change-Id: I525d43d1877cb38901c0725e4a2f4784a513bf69